Use consteval

+++ lib/include/Score.hpp
@@ -6,13 +6,13 @@
 
 namespace ms_pacman {
 
-constexpr int DEFAULT_LIVES = 3;
+constexpr uint16_t DEFAULT_LIVES = 3;
 constexpr int NORMAL_PELLET_POINTS = 10;
 constexpr int POWER_PELLET_POINTS = 50;
 constexpr int GHOST_POINTS = 200;
 
 struct Score {
-  int lives = DEFAULT_LIVES;
+  uint16_t lives = DEFAULT_LIVES;
   int points = 0;
   int eatenPellets = 0;
   std::vector<GenericFruit> eatenFruits;
